+++ src/test/java/io/github/astrapi69/json/JsonToYamlExtensionsTest.java
@@ -29,7 +29,6 @@ import static org.testng.AssertJUnit.assertEquals;
 import java.io.File;
 import java.io.IOException;
 
-import org.json.JSONException;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;