+++ src/test/java/io/github/astrapi69/json/factory/CustomJsonNodeFactoryTest.java
@@ -24,8 +24,6 @@
  */
 package io.github.astrapi69.json.factory;
 
-import static org.testng.Assert.assertNotNull;
-
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;